Added deletion by storage_id

+++ src/controllers/songController.ts
@@ -2,7 +2,7 @@
 import path from "path";
 import type { Request, Response } from "express";
 import { PrismaClient } from "@prisma/client";
-import { uploadSongToCloudinary } from "../services/cloudinaryService.js";
+import { deleteSongFromCloudinary, uploadSongToCloudinary } from "../services/cloudinaryService.js";
 import { durationToSeconds, getAudioDuration } from "./utilities.js";
 import { radioQueue } from "../services/RadioManager.js";
 
@@ -61,6 +61,7 @@ export const createSong = async (req: Request, res: Response) => {
       artist,
       duration: durationInSeconds,
       path: uploadResult.url!,
+      storage_id: uploadResult.public_id!,
     },
   });
 
@@ -78,7 +79,26 @@ export const updateSong = async (req: Request, res: Response) => {
 };
 
 export const deleteSong = async (req: Request, res: Response) => {
+  try {
   const { id } = req.params;
+  const song = await prisma.song.findUnique({
+    where: { id: Number(id) },
+    select: { storage_id: true },
+  });
+
+  if (!song) return res.json(404).json{error: "Song not found!"};
+
+  // Delete from cloudinary
+  if (song.storage_id) {
+    deleteSongFromCloudinary(song.storage_id);
+  }
+
+  // Delete from DB
   await prisma.song.delete({ where: { id: Number(id) } });
+  
   res.status(204).send();
+}catch (error) {
+  console.error("❌ Delete error", error);
+  return res.status(500).json({error: "Failed to delete song"});
+}
 };
